requirements.txt
streamlit>=1.28.0
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.1.0
faiss-cpu>=1.7.4
python-dotenv>=1.0.0
requests>=2.31.0
tavily-python>=0.2.0
pydantic>=2.0.0
ollama>=0.1.0
